pyproject.toml
# -*- coding: utf-8 -*-
#
# "TheVirtualBrain - Widgets" package
#
# (c) 2022-2025, TVB Widgets Team
#

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tvb-widgets"
readme = "README.md"
version = "2.2.2"
description = "TheVirtualBrain - Widgets is a package with somehow generic GUI components, but developed for EBRAINS Showcases in particular."
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
dependencies = [
    "colorcet",
    "ebrains_drive",
    "ipympl>0.8.5",
    "ipywidgets==8.0.7",
    "IPython<=8.12",
    "joblib",
    "mne>=1.0",
    "numpy<2",
    "plotly==5.14.0",
    "pythreejs",
    "k3d",
    "ipyreact>=0.4.1",
    "requests",
    "tvb-library",
    "tvb-framework",
    "plotly-resampler",
    "pyunicore>=1.0.0",
    "traitlets>=5.7.1",
    "toml",
    "bokeh",
    "ipycanvas",
    ]
keywords = ["tvb", "widgets", "jupyterlab", "ebrains", "components"]
authors = [
    { name = "TVB Widgets Team (Juelich SDL Neuroscience, INS - Marseille, Codemart)", email = "tvb.admin@thevirtualbrain.org" }
]

[project.urls]
Homepage = "https://github.com/the-virtual-brain/tvb-widgets"

[project.optional-dependencies]
examples = ["tvb-data"]
tests = ["pytest", "pytest-mock"]

[tool.hatch.build.targets.wheel]
include = [
    "tvbwidgets/",
    "codemeta.json"
]

[tool.hatch.build.targets.sdist]
include = [
    "tvbwidgets/",
    "codemeta.json"]

